General Overview
This is a COVID-19 Tracker in Tableau is a visual representation of the number of people who have been vaccinated against the virus, as well as the overall global population. It is designed to provide a clear and easily understandable overview of vaccination efforts around the world.

To calculate the maximum and minimum number of people who have been partially or fully vaccinated, the tracker uses a combination of calculations. One calculation would be to take the total number of people who have received at least one dose of a vaccine and divide that number by the overall population to find the percentage of people who have been partially vaccinated. Another calculation would be to take the total number of people who have received both doses of a vaccine and divide that by the overall population to find the percentage of people who are fully vaccinated.